feat: add layout (navbar and footer)
