+++ components/layout/Layout.jsx
@@ -1,0 +1,14 @@
+import Navbar from "./Navbar";
+import Footer from "./Footer";
+
+const Layout = ({ children }) => {
+    return (
+        <>
+            <Navbar />
+            <main>{children}</main>
+            <Footer />
+        </>
+    )
+}
+
+export default Layout;
