+++ components/layout/Navbar.jsx
@@ -1,0 +1,13 @@
+import Wrapper from "../Wrapper";
+
+const Navbar = () => {
+    return (
+        <Wrapper>
+            <nav>
+                Navigation Goes Here
+            </nav>
+        </Wrapper>
+    )
+}
+
+export default Navbar;
