+++ pages/index.js
@@ -1,6 +1,4 @@
 import Head from 'next/head'
-import Wrapper from '../components/Wrapper'
-import Footer from '../components/Footer'
 
 export default function Home() {
   return (
@@ -10,9 +8,6 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <Wrapper>
-        <Footer />
-      </Wrapper>
     </div>
   )
 }
